fixed the UTC encoding of the TimeEncoding.MAX_INSTANT

+++ yamcs-core/src/main/java/org/yamcs/utils/TaiUtcConverter.java
@@ -180,7 +180,7 @@ public class TaiUtcConverter {
     private long caldateToMjd(DateTimeComponents dtc) {
         int y;
         int m;
-        int d;
+        long d;
 
         d = dtc.day - 678882;
         m = dtc.month - 1;
